requirements/base.txt
# Archivo requirements base

django-bootstrap4==0.0.6
django-debug-toolbar==1.9.1
Django==2.0.7
djangorestframework==3.8.2
markdown2==2.3.5
Pillow==5.2.0
psycopg2-binary==2.7.5
pytz==2018.5
